requirements.txt
asgiref==3.8.1
Django==5.0.6
django-allauth==0.63.3
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-widget-tweaks==1.5.0
pillow==10.3.0
psycopg==3.1.19
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
